se agregaron las funciones que faltaban y se corrigieron errores de las funciones que ya estaban hechas. Falta testearlas a todas

+++ Funciones.h
@@ -41,5 +41,5 @@ void STOP(MV *mv);
 
 void errorMV(char *mensaje);
 
-void actualizarCC(MV *mv, long int resultado, int carry, int overflow);
+void actualizarCC(MV *mv, int32_t resultado, int carry, int overflow);
 #endif
